fix: move clone form to own component
